SVN trunk(2.2) commit: Bug#324341 cascade delete support
https://bugs.eclipse.org/bugs/show_bug.cgi?id=324341

http://wiki.eclipse.org/EclipseLink/DesignDocs/324341

Code review: Andrei (pending), Guy

Changes:
- Changed Project.getOrderedDescriptors() to return List instead of Vector, updated dbws and moxy references.
- Removed some dead Helper methods and tests.
- Added CascadeOnDelete annotation.
- Added isCascadeOnDeleteSetOnDatabaseOnSecondaryTables to ClassDescriptor.
- Added objectIsEasilyInstantiated to indirection policy.
- Improved logging to log on a flush, and log client session id when acquired.
- Added cascadeOnDelete support to DeleteObject and DeleteAll query mechanism prepare.
- Added cascadeDeleteObjects set to UnitOfWork to track cascade deletion.
- Added delete cascade support to mappings, fixed deletion to not instantiate collection if not required, not execute delete if empty.
- Changed mustDeleteReferenceObjectsOneByOne to an instance variable instead of method, compute in init.
- Changed verifyDelete to also check cascade remove mappings.
- Changed hasDependency to check cascade remove to process delete-all correctly.
- Changed cascade remove to not instantiate the relationship if delete-all can be used when private owned.
- Fixed m-m mapping to still delete if private in a unit of work (was recently broken).
- Fixed Unidirectional 1-m to not cascade delete to map keys if not private owned. 
- Added in memory only option to DeleteAll query to just update the cache.
- Changed DefaultTableGenerator to generate cascade delete in foreign keys if set, added foreign key constraints to direct collection and aggregate collection.
- Added CascadeOnDelete annotation processing to JPA metadata, refactored common relationship processing.
- Added some @Overrides, removed some dead code, change some Vector to generic Lists, some cleanup/micro.

Testing:
- Added cascade delete to JPA DDL test model.
- Added cascade delete to JPA fieldaccess employee test model.
- Added delete tests to fieldaccess entity manager test suite.

+++ foundation/org.eclipse.persistence.core/src/org/eclipse/persistence/queries/DeleteObjectQuery.java
@@ -133,7 +133,7 @@ public class DeleteObjectQuery extends ObjectLevelModifyQuery {
         Object object = getObject();
         boolean isUnitOfWork = session.isUnitOfWork();
         try {
-            // Check if the object has already been commited, then no work is required
+            // Check if the object has already been committed, then no work is required
             if (commitManager.isProcessedCommit(object)) {
                 return object;
             }
@@ -165,7 +165,15 @@ public class DeleteObjectQuery extends ObjectLevelModifyQuery {
             if (QueryMonitor.shouldMonitor()) {
                 QueryMonitor.incrementDelete(this);
             }
-            int rowCount = getQueryMechanism().deleteObject().intValue();
+            int rowCount = 0;
+            // If the object was/will be deleted from a cascade delete constraint, ignore it.
+            if (isUnitOfWork && ((UnitOfWorkImpl)session).hasCascadeDeleteObjects()
+                    && ((UnitOfWorkImpl)session).getCascadeDeleteObjects().contains(object)) {
+                // Cascade delete does not check optimistic lock, assume ok.
+                rowCount = 1;
+            } else {
+                rowCount = getQueryMechanism().deleteObject().intValue();
+            }
 
             if (rowCount < 1) {
                 if (session.hasEventManager()) {
